requirements.txt
vt-py
python-dotenv
opencv-python
pyzbar
numpy